src/backend/requirements.txt
bcrypt
dotenv
flask
flask-cors
mysql-connector-python
requests